feat(ui): abstracting renderer
