+++ src/view/index.ts
@@ -3,8 +3,9 @@ import { getAPI, isPluginEnabled } from "obsidian-dataview";
 
 import ObsidianFeedsPlugin from "main";
 import getSettings, { Settings } from "view/settings";
+import FeedsRenderer from "view/view";
 import { renderError } from "ui/render";
-import { FeedsRenderer } from "ui/feeds-renderer";
+import { RefreshableRenderer } from "ui/renderer";
 
 export default class ObsidianFeeds extends MarkdownRenderChild {
   private settings: Settings;
@@ -17,7 +18,7 @@ export default class ObsidianFeeds extends MarkdownRenderChild {
     public ctx: MarkdownPostProcessorContext,
   ) {
     super(containerEl);
-    this.settings = getSettings(src, app, containerEl, ctx);
+    this.settings = getSettings(src, containerEl, app, ctx);
   }
 
   async onload() {
@@ -28,11 +29,10 @@ export default class ObsidianFeeds extends MarkdownRenderChild {
       renderError(this.containerEl, error);
       throw new Error(error);
     }
-    const dv = getAPI();
 
-    console.log({ dv });
-
-    this.addChild(new FeedsRenderer(dv, this.containerEl, this.ctx.sourcePath));
+    const dvApi = getAPI();
+    const feedsRenderer = new FeedsRenderer(dvApi, this.containerEl, this.ctx, this);
+    this.addChild(new RefreshableRenderer(dvApi, this.containerEl, feedsRenderer));
   }
 
   async onunload() {}
